refactor: step4 피드백 반영- winners 분리
